requirements.txt
requests
SQLAlchemy
python-dateutil
numpy
APScheduler
python-dotenv
nltk
pytrends
